requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
Flask==3.0.3
Flask_Cors==4.0.0
flask_jwt_extended==4.6.0
Flask_RESTful==0.3.10
google_api_python_client==2.136.0
google_auth_oauthlib==1.2.1
openmeteo_requests==1.3.0
pandas==2.2.3
protobuf==5.28.2
PyJWT==2.9.0
PyJWT==2.9.0
pymongo==4.10.1
python-dotenv==1.0.1
pytz==2024.1
Requests==2.32.3
requests_cache==1.2.1
retry_requests==2.0.0
rich==13.9.2
SQLAlchemy==2.0.35
tenacity==9.0.0
websocket_client==1.8.0
Werkzeug==3.0.4
yfinance==0.2.37
psycopg2-binary==2.9.9
gunicorn==23.0.0
Flask-Caching==1.10.1
premailer==3.10.0
Flask-Limiter==3.5.0